starting adding couriers

+++ Task_2_2_1/src/main/java/nsu/fit/tsukanov/storage/Storage.java
@@ -16,5 +16,5 @@ public interface Storage {
 
     Order takePizzaOrder();
 
-    List<Order> takePizzaOrders(long amount);
+    List<Order> takePizzaOrders(int amount);
 }
